Cleanup and refactorings
