+++ src/main/kotlin/no/skatteetaten/aurora/boober/service/DeployService.kt
@@ -52,7 +52,7 @@ class DeployService(
     }
 
 
-    private fun prepareDeployEnvironments(deploymentSpecs: List<AuroraDeploymentSpec>): Map<AuroraDeployEnvironment, AuroraDeployResult> {
+    fun prepareDeployEnvironments(deploymentSpecs: List<AuroraDeploymentSpec>): Map<AuroraDeployEnvironment, AuroraDeployResult> {
 
         val authenticatedUser = userDetailsProvider.getAuthenticatedUser()
 
@@ -85,7 +85,7 @@ class DeployService(
     }
 
 
-    private fun prepareDeployEnvironment(environment: AuroraDeployEnvironment, projectExist: Boolean): List<OpenShiftResponse> {
+    fun prepareDeployEnvironment(environment: AuroraDeployEnvironment, projectExist: Boolean): List<OpenShiftResponse> {
 
         val affiliation = environment.affiliation
         val namespace = environment.namespace
